option trait cleanup
