+++ gateware/src/example_soc/fw/src/lib.rs
@@ -17,3 +17,4 @@ tiliqua_hal::impl_i2c! {
 }
 
 pub mod log;
+pub mod opts;
